requirements.txt
streamlit
openai-whisper
openai 